Compiles again

+++ liballoc.h
@@ -2,9 +2,9 @@
 #define _LIBALLOC_H
 
 
-/** \defgroup ALLOCHOOKS liballoc hooks 
+/** \defgroup ALLOCHOOKS liballoc hooks
  *
- * These are the OS specific functions which need to 
+ * These are the OS specific functions which need to
  * be implemented on any platform that the library
  * is expected to work on.
  */
@@ -23,7 +23,7 @@ typedef	unsigned int	size_t;
 
 #ifndef _HAVE_UINTPTR_T
 #define _HAVE_UINTPTR_T
-typedef	unsigned int	uintptr_t;
+typedef	unsigned long long	uintptr_t;
 #endif
 
 
@@ -44,7 +44,7 @@ extern "C" {
 
 /** This function is supposed to lock the memory data structures. It
  * could be as simple as disabling interrupts or acquiring a spinlock.
- * It's up to you to decide. 
+ * It's up to you to decide.
  *
  * \return 0 if the lock was acquired successfully. Anything else is
  * failure.
@@ -79,7 +79,7 @@ extern void* liballoc_alloc(int);
 extern int liballoc_free(void*,int);
 
 
-       
+
 
 void     *PREFIX(malloc)(size_t);				///< The standard function.
 void     *PREFIX(realloc)(void *, size_t);		///< The standard function.
